extract postprocess method in AposEditor composable, use it in AposEditorMixin, use with quick upload
