+++ modules/@apostrophecms/modal/ui/apos/mixins/AposEditorMixin.js
@@ -18,6 +18,7 @@ import { klona } from 'klona';
 import {
   evaluateExternalConditions, getConditionalFields, getConditionTypesObject
 } from 'Modules/@apostrophecms/schema/lib/conditionalFields.js';
+import { _postprocess } from '../composables/AposEditor.js';
 
 export default {
   props: {
@@ -219,57 +220,12 @@ export default {
         this.triggerValidation = false;
       });
     },
-    // Perform any postprocessing required by direct or nested schema fields
-    // before the object can be saved
     async postprocess() {
-      // Relationship fields may have postprocessors (e.g. autocropping)
-      const relationships = findRelationships(this.schema, this.docFields.data);
-      for (const relationship of relationships) {
-        if (!(relationship.value && relationship.field.postprocessor)) {
-          continue;
-        }
-        const withType = relationship.field.withType;
-        const mod = apos.modules[withType];
-        const response = await apos.http.post(`${mod.action}/${relationship.field.postprocessor}`, {
-          qs: {
-            aposMode: 'draft'
-          },
-          body: {
-            relationship: relationship.value,
-            // Pass the options of the widget currently being edited, some
-            // postprocessors need these
-            // (e.g. autocropping cares about widget aspectRatio)
-            widgetOptions: apos.area.widgetOptions[0]
-          },
-          busy: true
-        });
-        relationship.context[relationship.field.name] = response.relationship;
-      }
-      function findRelationships(schema, object) {
-        let relationships = [];
-        for (const field of schema) {
-          if (field.type === 'relationship') {
-            relationships.push({
-              context: object,
-              field,
-              value: object[field.name]
-            });
-          } else if (field.type === 'array') {
-            for (const value of (object[field.name] || [])) {
-              relationships = [
-                ...relationships,
-                findRelationships(field.schema, value)
-              ];
-            }
-          } else if (field.type === 'object') {
-            relationships = [
-              ...relationships,
-              findRelationships(field.schema, object[field.name] || {})
-            ];
-          }
-        }
-        return relationships;
-      }
+      await _postprocess(
+        this.schema,
+        this.docFields.data,
+        apos.area.widgetOptions[0]
+      );
     }
   }
 };
